pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kbi-labs"
version = "2.0.0"
description = "KBI Labs Intelligence Platform for SMB Government Contractors"
authors = [{name = "KBI Labs", email = "support@kbilabs.com"}]
license = {text = "Proprietary"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
    "sqlalchemy>=2.0.42",
    "asyncpg>=0.29.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.0.0",
    "passlib[bcrypt]>=1.7.4",
    "PyJWT>=2.8.0",
    "slowapi>=0.1.9",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-v",
    "--tb=short",
    "--cov=.",
    "--cov-report=html:htmlcov",
    "--cov-report=term-missing",
    "--cov-fail-under=90"
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "venv/*",
    "*/__pycache__/*",
    "*/migrations/*",
    "setup.py",
    "conftest.py"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:"
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | venv
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true

[tool.flake8]
max-line-length = 100
exclude = [
    ".git",
    "__pycache__",
    "venv",
    ".venv",
    "build",
    "dist"
]
ignore = [
    "E203",  # Whitespace before ':'
    "W503",  # Line break before binary operator
    "E501"   # Line too long (handled by black)
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "asyncpg.*",
    "aiosqlite.*",
    "slowapi.*"
]
ignore_missing_imports = true

[tool.bandit]
targets = ["main_server.py", "backend/"]
exclude_dirs = ["tests", "venv", ".git"]
severity = "medium"
confidence = "medium"

[tool.safety]
# Safety dependency vulnerability scanner
full_report = true
output = "json"